specializer.py: Factor out "special_instruction.py"
This is so that it can be used in gen-labels.
